separate helper functions from app logic
