+++ app.py
@@ -2,6 +2,8 @@ import streamlit as st
 import pandas as pd
 from typing import Dict, Optional
 
+from helper import *
+
 # State Initialization
 def initialize_state():
     """Initializes or resets the session state variables."""
@@ -13,110 +15,6 @@ def initialize_state():
 if 'accounts' not in st.session_state:
     initialize_state()
 
-# Helper Functions
-
-def load_transaction_file(uploaded_file) -> Optional[pd.DataFrame]:
-    """
-    Loads and parses the uploaded CSV file into a DataFrame without headers.
-    Assigns appropriate column names.
-    Returns None if an error occurs.
-    """
-    try:
-        # Define the column names
-        column_names = [
-            'Account Name',
-            'Card Number',
-            'Transaction Amount',
-            'Transaction Type',
-            'Description',
-            'Target Card Number'
-        ]
-        data = pd.read_csv(uploaded_file, names=column_names, header=None)
-        return data
-    except Exception as e:
-        st.error(f"Error reading the file: {e}")
-        return None
-
-
-def validate_transaction(row: pd.Series) -> Optional[str]:
-    """
-    Validates a single transaction row.
-    Returns an error message if validation fails, otherwise None.
-    """
-    required_fields = [
-        'Account Name',
-        'Card Number',
-        'Transaction Amount',
-        'Transaction Type',
-        'Description'
-    ]
-    for field in required_fields:
-        if pd.isna(row.get(field)) or row.get(field) == '':
-            return f"Missing required field: {field}"
-
-    transaction_type = str(row['Transaction Type']).strip()
-    if transaction_type not in ['Credit', 'Debit', 'Transfer']:
-        return f"Invalid Transaction Type: {transaction_type}"
-    
-    amount = row.get("Transaction Amount")
-    try:
-        _ = float(amount)
-    except ValueError:
-        return "Invalid Transaction Amount format"
-
-    if pd.isna(row.get('Card Number')) or row.get('Card Number') == '':
-        return "Missing Card Number"
-    else:
-        # Ensure Card Number is valid
-        try:
-            _ = str(int(float(row['Card Number'])))
-        except ValueError:
-            return "Invalid Card Number format"
-
-    if transaction_type == 'Transfer':
-        if pd.isna(row.get('Target Card Number')) or row.get('Target Card Number') == '':
-            return "Missing Target Card Number for Transfer"
-        else:
-            # Ensure Target Card Number is valid
-            try:
-                _ = str(int(float(row['Target Card Number'])))
-            except ValueError:
-                return "Invalid Target Card Number format"
-    return None  # No validation errors
-
-def process_transaction(row: pd.Series, accounts: Dict[str, Dict[str, float]], card_to_account: Dict[str, str]):
-    """
-    Processes a single valid transaction and updates the accounts dictionary.
-    """
-    account_name = str(row['Account Name']).strip()
-    card_number = str(int(float(row['Card Number'])))
-    transaction_amount = float(row['Transaction Amount'])
-    transaction_type = str(row['Transaction Type']).strip()
-    description = str(row['Description']).strip()
-    target_card_number = row.get('Target Card Number')
-
-    card_to_account[card_number] = account_name
-    # Ensure the account and card exist in the accounts dictionary
-    accounts.setdefault(account_name, {}).setdefault(card_number, 0.0)
-
-    if transaction_type == 'Credit':
-        accounts[account_name][card_number] += transaction_amount
-    elif transaction_type == 'Debit':
-        accounts[account_name][card_number] -= transaction_amount
-    elif transaction_type == 'Transfer':
-        target_card_number = str(int(float(target_card_number)))
-        target_account_name = card_to_account.get(target_card_number)
-
-        if target_account_name is None:
-            return False
-        
-        accounts[target_account_name].setdefault(target_card_number, 0.0)
-  
-        accounts[account_name][card_number] -= transaction_amount
-        accounts[target_account_name][target_card_number] += transaction_amount
-    
-    return True
-
 
 def process_transactions(data: pd.DataFrame):
     """
@@ -127,6 +25,7 @@ def process_transactions(data: pd.DataFrame):
     card_to_account = st.session_state['card_to_account']
     deferred_transfers = st.session_state['deferred_transfers']
 
+    # Process all non-transfer transactions
     for _, row in data.iterrows():
         error_message = validate_transaction(row)
         if error_message:
@@ -135,11 +34,7 @@ def process_transactions(data: pd.DataFrame):
         else:
             transaction_type = str(row['Transaction Type']).strip()
             if transaction_type != 'Transfer':
-                try:
-                    _ = process_transaction(row, accounts, card_to_account)
-                except Exception as e:
-                    row['Error'] = f"Processing error: {e}"
-                    st.session_state['bad_transactions'] = st.session_state['bad_transactions']._append(row, ignore_index=True)
+                _ = process_transaction(row, accounts, card_to_account)
             else:
                 deferred_transfers.append(row)
     
@@ -164,7 +59,7 @@ def process_transactions(data: pd.DataFrame):
     # remaining deffered_transfers are bad transactions
     for row in deferred_transfers:
         row['Error'] = "Unkown target card number"
-        st.session_state['bad_transactions'] = st.session_state['bad_transactions'].append(row, ignore_index=True)
+        st.session_state['bad_transactions'] = st.session_state['bad_transactions']._append(row, ignore_index=True)
 
 
 
@@ -198,18 +93,19 @@ def display_dataframe(title: str, df: pd.DataFrame):
     else:
         st.write(f"No data to display for {title.lower()}.")
 
-# Main Application Logic
 
 def main():
     st.title("Transaction Processor")
     st.write("Upload a CSV file containing transactions to process them.")
 
     # Sidebar for file upload and reset button
+    uploaded_file = None
     with st.sidebar:
         st.header("Options")
         uploaded_file = st.file_uploader("Choose a CSV file", type="csv")
         if st.button("Reset System"):
             initialize_state()
+            uploaded_file = None
             st.success("System has been reset.")
 
     # Process the uploaded file
